requirements.txt
gevent==0.13.8
gevent-socketio==0.3.5-rc2
gevent-websocket==0.3.6
gevent-zeromq==0.2.5
pyzmq==2.2.0
